+++ client/capi/base/autd3_c_api.h
@@ -3,7 +3,7 @@
 // Created Date: 07/02/2018
 // Author: Shun Suzuki
 // -----
-// Last Modified: 10/08/2021
+// Last Modified: 03/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2018-2020 Hapis Lab. All rights reserved.
@@ -76,8 +76,8 @@ EXPORT_AUTD void AUTDDeleteModulation(void* mod);
 
 EXPORT_AUTD void AUTDSequence(void** out);
 EXPORT_AUTD void AUTDGainSequence(void** out, uint16_t gain_mode);
-EXPORT_AUTD bool AUTDSequenceAddPoint(void* seq, double x, double y, double z);
-EXPORT_AUTD bool AUTDSequenceAddPoints(void* seq, const double* points, uint64_t size);
+EXPORT_AUTD bool AUTDSequenceAddPoint(void* seq, double x, double y, double z, uint8_t duty);
+EXPORT_AUTD bool AUTDSequenceAddPoints(void* seq, const double* points, uint64_t points_size, uint8_t* duties, uint64_t duties_size);
 EXPORT_AUTD bool AUTDSequenceAddGain(void* seq, void* gain);
 EXPORT_AUTD double AUTDSequenceSetFreq(void* seq, double freq);
 EXPORT_AUTD double AUTDSequenceFreq(void* seq);