+++ client/include/autd3/core/ec_config.hpp
@@ -3,7 +3,7 @@
 // Created Date: 05/11/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 04/07/2021
+// Last Modified: 03/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -25,9 +25,6 @@ constexpr uint32_t EC_SYNC0_CYCLE_TIME_MICRO_SEC = 500;
 constexpr uint32_t EC_SM3_CYCLE_TIME_NANO_SEC = EC_SM3_CYCLE_TIME_MICRO_SEC * 1000;
 constexpr uint32_t EC_SYNC0_CYCLE_TIME_NANO_SEC = EC_SYNC0_CYCLE_TIME_MICRO_SEC * 1000;
 
-constexpr uint8_t READ_MOD_IDX_BASE_HEADER = 0xC0;
-constexpr uint8_t READ_MOD_IDX_BASE_HEADER_AFTER_SHIFT = 0xE0;
-constexpr uint8_t READ_MOD_IDX_BASE_HEADER_MASK = 0xE0;
 constexpr size_t EC_DEVICE_PER_FRAME = 2;
 constexpr size_t EC_FRAME_LENGTH = 14 + 2 + (10 + EC_OUTPUT_FRAME_SIZE + EC_INPUT_FRAME_SIZE + 2) * EC_DEVICE_PER_FRAME + 10;
 constexpr double EC_SPEED_BPS = 100.0 * 1000 * 1000;