+++ client/include/autd3/sequence/primitive.hpp
@@ -3,7 +3,7 @@
 // Created Date: 14/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 28/07/2021
+// Last Modified: 03/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -32,7 +32,8 @@ class Circumference : PointSequence {
    * @param[in] normal Normal vector of the circumference
    * @param[in] radius Radius of the circumference
    * @param[in] n Number of the control points
+   * @param[in] duty duty of control points
    */
-  static PointSequencePtr create(const core::Vector3& center, const core::Vector3& normal, double radius, size_t n);
+  static PointSequencePtr create(const core::Vector3& center, const core::Vector3& normal, double radius, size_t n, uint8_t duty = 0xFF);
 };
 }  // namespace autd::sequence