+++ client/lib/controller.cpp
@@ -3,7 +3,7 @@
 // Created Date: 05/11/2020
 // Author: Shun Suzuki
 // -----
-// Last Modified: 13/08/2021
+// Last Modified: 03/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -281,6 +281,24 @@ void Controller::STMController::finish() {
 
 void Controller::STMController::stop() {
   if (this->_handler == nullptr) this->_handler = this->_timer->stop();
+}
+
+void Controller::STMTimerCallback::add(std::unique_ptr<uint8_t[]> data, const size_t size) {
+  this->_bodies.emplace_back(std::move(data));
+  this->_sizes.emplace_back(size);
+}
+void Controller::STMTimerCallback::clear() {
+  std::vector<std::unique_ptr<uint8_t[]>>().swap(this->_bodies);
+  std::vector<size_t>().swap(this->_sizes);
+  this->_idx = 0;
+}
+
+void Controller::STMTimerCallback::callback() {
+  if (auto expected = false; _lock.compare_exchange_weak(expected, true)) {
+    this->_link->send(&this->_bodies[_idx][0], this->_sizes[_idx]);
+    this->_idx = (this->_idx + 1) % this->_bodies.size();
+    _lock.store(false, std::memory_order_release);
+  }
 }
 
 }  // namespace autd