+++ client/lib/sequence/primitive.cpp
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 20/07/2021
+// Last Modified: 03/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -15,7 +15,8 @@
 
 namespace autd::sequence {
 
-PointSequencePtr Circumference::create(const core::Vector3& center, const core::Vector3& normal, const double radius, const size_t n) {
+PointSequencePtr Circumference::create(const core::Vector3& center, const core::Vector3& normal, const double radius, const size_t n,
+                                       const uint8_t duty) {
   auto get_orthogonal = [](const core::Vector3& v) {
     const auto a = core::Vector3::UnitX();
     if (std::acos(v.dot(a)) < M_PI / 2) return v.cross(core::Vector3::UnitY());
@@ -26,13 +27,13 @@ PointSequencePtr Circumference::create(const core::Vector3& center, const core::
   const auto n1 = get_orthogonal(normal_).normalized();
   const auto n2 = normal_.cross(n1).normalized();
 
-  std::vector<core::Vector3> control_points;
+  auto seq = PointSequence::create();
   for (size_t i = 0; i < n; i++) {
     const auto theta = 2 * M_PI / static_cast<double>(n) * static_cast<double>(i);
     auto x = n1 * radius * std::cos(theta);
     auto y = n2 * radius * std::sin(theta);
-    control_points.emplace_back(center + x + y);
+    seq->add_point(center + x + y, duty);
   }
-  return PointSequence::create(control_points);
+  return seq;
 }
 }  // namespace autd::sequence